requirements.txt
# Telegram bot
aiogram==3.14.0

# Database
asyncpg==0.30.0

# Blackbird core dependencies
aiohttp>=3.9,<3.11
requests==2.32.4
rich==14.0.0
python-dotenv==1.1.1
pillow==11.3.0
reportlab==4.4.2

# Utils
chardet==5.2.0
certifi==2025.6.15

# Email enrichment
dnspython>=2.4.0
python-whois>=0.9.0
httpx>=0.27.0
trio>=0.25.0
tqdm>=4.66.0
termcolor>=2.4.0
bs4>=0.0.2
colorama>=0.4.6
